Migrate to using Cobra for a nicer CLI and future possibilities
