+++ .gitignore
@@ -5,7 +5,7 @@
 *.so
 *.dylib
 
-server
+bearcatter
 
 # Test binary, built with `go test -c`
 *.test
