--- go.mod
@@ -1,17 +1,0 @@
-module github.com/Bearcatter/bearcatter
-
-go 1.14
-
-require (
-	github.com/davecgh/go-spew v1.1.1
-	github.com/go-audio/riff v1.0.0
-	github.com/go-playground/validator/v10 v10.3.0
-	github.com/gobwas/httphead v0.0.0-20180130184737-2c6c146eadee // indirect
-	github.com/gobwas/pool v0.2.0 // indirect
-	github.com/gobwas/ws v1.0.3
-	github.com/gocarina/gocsv v0.0.0-20200330101823-46266ca37bd3
-	github.com/sirupsen/logrus v1.6.0
-	github.com/spf13/pflag v1.0.5
-	github.com/stretchr/testify v1.4.0
-	github.com/tarm/serial v0.0.0-20180830185346-98f6abe2eb07
-)
