+++ server/structs.go
@@ -1,4 +1,4 @@
-package main
+package server
 
 import (
 	"encoding/xml"
@@ -8,6 +8,86 @@ import (
 	"time"
 
 	"github.com/Bearcatter/bearcatter/wavparse"
+)
+
+type SDSKeyType string
+type SDSKeyModeType string
+type GltXmlType int
+
+const (
+	DefaultGoProcMultiplier = 5
+	DefaultGoProcDelay      = 30 // milliseconds
+
+	LF = 1 << 0
+	NL = 1 << 1
+
+	VALID_KEY_CMD_LENGTH = 10
+	KEY_PUSH_CMD         = "PUSH"
+	KEY_HOLD_CMD         = "HOLD"
+
+	KEY_MENU      SDSKeyType = "M"
+	KEY_F         SDSKeyType = "F"
+	KEY_1         SDSKeyType = "1"
+	KEY_2         SDSKeyType = "2"
+	KEY_3         SDSKeyType = "3"
+	KEY_4         SDSKeyType = "4"
+	KEY_5         SDSKeyType = "5"
+	KEY_6         SDSKeyType = "6"
+	KEY_7         SDSKeyType = "7"
+	KEY_8         SDSKeyType = "8"
+	KEY_9         SDSKeyType = "9"
+	KEY_0         SDSKeyType = "0"
+	KEY_DOT       SDSKeyType = "."
+	KEY_ENTER     SDSKeyType = "E"
+	KEY_ROT_RIGHT SDSKeyType = ">"
+	KEY_ROT_LEFT  SDSKeyType = "<"
+	KEY_ROT_PUSH  SDSKeyType = "^"
+	KEY_VOL_PUSH  SDSKeyType = "V"
+	KEY_SQL_PUSH  SDSKeyType = "Q"
+	KEY_REPLAY    SDSKeyType = "Y"
+	KEY_SYSTEM    SDSKeyType = "A"
+	KEY_DEPT      SDSKeyType = "B"
+	KEY_CHANNEL   SDSKeyType = "C"
+	KEY_ZIP       SDSKeyType = "Z"
+	KEY_SERV      SDSKeyType = "T"
+	KEY_RANGE     SDSKeyType = "R"
+
+	KEY_MODE_PRESS   SDSKeyModeType = "P" // Press (One Push)
+	KEY_MODE_LONG    SDSKeyModeType = "L" // Long Press (Press and Hold a few seconds)
+	KEY_MODE_HOLD    SDSKeyModeType = "H" // Hold (Press and Hold until Release receive)
+	KEY_MODE_RELEASE SDSKeyModeType = "R" // Release (Cancel Hold state
+
+	GltXmlUnknown GltXmlType = -1
+	GltXmlFL      GltXmlType = iota
+	GltXmlSYS
+	GltXmlDEPT
+	GltXmlSITE
+	GltXmlCFREQ
+	GltXmlTGID
+	GltXmlSFREQ
+	GltXmlAFREQ
+	GltXmlATGID
+	GltXmlFTO
+	GltXmlCSBANK
+	GltXmlUREC
+	GltXmlIREC_FILE
+	GltXmlUREC_FOLDER
+	GltXmlUREC_FILE
+	GltXmlTRN_DISCOV
+	GltXmlCNV_DISCOV
+
+	AstModeCurrentActivity ASTModeType = "CURRENT_ACTIVITY"
+	AstModeLCNMonitor                  = "LCN_MONITOR"
+	AstActivityLog                     = "ACTIVITY_LOG"
+	AstLCNFinder                       = "LCN_FINDER"
+
+	AprModePause APRModeType = "PAUSE"
+	AprModeRESME APRModeType = "RESUME"
+)
+
+var (
+	validKeys = loadValidKeys()
+	TERMINATE = "quit\r"
 )
 
 type ScannerInfo struct {
